+++ app/src/main/java/com/example/andrhomeworks/OnClickButton.kt
@@ -4,6 +4,6 @@ interface OnClickButton {
     fun onClick()
     fun addEmployee()
     fun initDataFragment(id:Long)
-    fun editUser(id: Long)
-    fun deleteUser(id: Long)
+    fun editUser(id: Int)
+    fun deleteUser(id: Int)
 }
